move DependencyTests
