+++ tests/IntegrationTests/Curiosity.Migrations.IntegrationTests/Curiosity.Migrations.IntegrationTests.csproj
@@ -5,7 +5,7 @@
 
         <IsPackable>false</IsPackable>
 
-        <RootNamespace>Curiosity.Migrations.TransactionTests</RootNamespace>
+        <RootNamespace>Curiosity.Migrations.IntegrationTests</RootNamespace>
 
         <LangVersion>11</LangVersion>
 
@@ -52,6 +52,15 @@
         <CopyToOutputDirectory>Always</CopyToOutputDirectory>
       </None>
       <None Update="TransactionScriptMigrations\6.0-incorrect_script.sql">
+        <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+      </None>
+      <None Update="DependencyScriptMigrations\1.0-correct_script.sql">
+        <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+      </None>
+      <None Update="DependencyScriptMigrations\4.0-script_with_ok_dependencies.sql">
+        <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+      </None>
+      <None Update="DependencyScriptMigrations\7.0-script_with_not_ok_dependencies.sql">
         <CopyToOutputDirectory>Always</CopyToOutputDirectory>
       </None>
     </ItemGroup>
