+++ tests/IntegrationTests/Curiosity.Migrations.IntegrationTests/DependencyCodeMigrations/IDependencyMigration.cs
@@ -1,0 +1,6 @@
+﻿namespace Curiosity.Migrations.IntegrationTests.DependencyCodeMigrations;
+
+public interface IDependencyMigration
+{
+    
+}
