+++ tests/IntegrationTests/Curiosity.Migrations.IntegrationTests/DependencyScriptMigrations/1.0-correct_script.sql
