+++ tests/IntegrationTests/Curiosity.Migrations.IntegrationTests/DependencyScriptMigrations/4.0-script_with_ok_dependencies.sql
