+++ tests/IntegrationTests/Curiosity.Migrations.IntegrationTests/DependencyScriptMigrations/7.0-script_with_not_ok_dependencies.sql
