+++ tests/IntegrationTests/Curiosity.Migrations.IntegrationTests/TransactionCodeMigrations/CodeMigration_2_0.cs
@@ -2,7 +2,7 @@ using System.Data.Common;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Curiosity.Migrations.TransactionTests.TransactionCodeMigrations;
+namespace Curiosity.Migrations.IntegrationTests.TransactionCodeMigrations;
 
 public class CodeMigration_2_0 : CodeMigration, ITransactionMigration
 {
