+++ tests/IntegrationTests/Curiosity.Migrations.IntegrationTests/TransactionCodeMigrations/ITransactionMigration.cs
@@ -1,4 +1,4 @@
-﻿namespace Curiosity.Migrations.TransactionTests.TransactionCodeMigrations;
+﻿namespace Curiosity.Migrations.IntegrationTests.TransactionCodeMigrations;
 
 public interface ITransactionMigration
 {
